feat: added sign-up/login buttons to menu
